allsoc: docs: update readme for repo tool installation
If one follows the repo tool instructions from the README, it might not
work out of box owing to some PATH and executable permission issues from
inside of GitBash in case of Windows host.  Update the README and give
clear instructions to setup repo tool for OOB usage in windows.

Fixes: MCUSDK-9099

Signed-off-by: Anand Mahadevan SS <a0230421@ti.com>