allsoc: syscfg: ospi: fix load-from-json button
The load-from-json button in flash sysconfig which allows to import
flash configuration from a browsed JSON file doesn't work in windows.
This is probably because of the spawn method in child_process not
identifying the `copy` command in windows. To remove the ambiguity use a
JS based copy util script.

Fixes: MCUSDK-8383

Signed-off-by: Anand Mahadevan SS <a0230421@ti.com>