am263x: epwm: Add mixed events source selection in SYSCFG for DC and ET submodule
When we select "Time-base counter based on mixed events" for starting Blanking window in Digital Compare module,
or for selecting the source to trigger interrupt or as ADC SOC in Event Trigger, we don't get to see
any option of choosing what all mixed events do we want. Add the list of mixed events for the user to choose from.

Fixes: MCUSDK-9313

Signed-off-by: Sudeshna Chaudhuri <s-chaudhuri@ti.com>