am263x: watchdog: Add user configurability for clock frequencies
Currently Watchdog Syscfg doesn't give the option to choose clock sources and
input clock frequencies. It's hardcoded to SYSCLK and 5MHz respectively.
This is an invalid combination for WDTx_CLK_DIV_VAL.
WDTx_CLK_DIV_VAL can take values from 0 to 15 but SYSCLK/5MHz = 40.
Add the necessary support for this.

Fixes: MCUSDK-8989

Signed-off-by: Sudeshna Chaudhuri <s-chaudhuri@ti.com>