allsoc: uniflash: fix 1MB file bug in uart_uniflash
When files with size > 1MB are sent for flashing in UART Uniflash, it
splits it into chunks of 1MB and sends it. There is a mistake in
calculating the residual bytes after the 1 MB chunks are sent, because
of this the image is flashed incorrectly. Fix this.

Also add > 1MB test apps to test the UART Uniflash.

Fixes: MCUSDK-8414
Signed-off-by: Anand Mahadevan SS <a.mahadevanss@ti.com>