am243x/am263x: uart: fix baudrate restraints
UART syscfg for am243x and am263x restricts higher baudrates with 16x
operational mode. From experimentation this is not applicable for higher
baudrates. Modify the validation check to include this.

Also fix the bug that the inputClk value used to compute divisor is
actually the hardcoded 48 MHz.

Update the UART syscfg for both SoCs in general for more configurability

Fixes: MCUSDK-8403
Signed-off-by: Anand Mahadevan SS <a.mahadevanss@ti.com>