am64x-am243x: USB: SBL: dfu-util timeout resolve
- Resolve dfu-util timeout error. We need some way
to poll the current DFU state
The static global variable exists in .bss section.
we get the start address of that section from linker
script and poll the status.
whenever a condition
( MANIFEST_DONE && DFU state returned to IDLE )
is true then exit the dfu_task and boot the received appimage. 

Fixes: MCUSDK-9821

Signed-off-by: Keval Joshi <k-joshi@ti.com>