am263x: bootloader: Bootloader fix for HS-FS PG1A
change default retVal to SystemP_SUCCESS as retVal can only be a SystemP_FAILURE
when Hsmclient_loadHSMRtFirmware returns SystemP_FAILURE in case of only HS-FS
PG1.1 and HSSE device.

The default value should be SystemP_SUCCESS as assert will be blocking even if
there is no need of loading HSMrt.

Fixes: MCUSDK-9876

Signed-off-by: Utkarsh M Kripashankar <u-kripashankar@ti.com>