am263x: uniflash: uart: fix timeout error
UART Uniflash errors out while flashing images > 500 KB.
It seems to be an xmodem receive timeout issue. Fix this
by increasing the serial port timeout in host side scripts.
Also provide an option in the CLI script to specify size of
the receiving buffer in case it is lesser than 1MB. Also
cleanup the target side code a bit for robustness.

Fixes: MCUSDK-9890
Signed-off-by: Anand Mahadevan SS <a.mahadevanss@ti.com>