am64x/am243x: sbl: give option to disable app auth
SYSFW has a bug in in-place authentication because of which it takes
a long time to authenticate appimages. This affects the boot time. So
provide an option via sysconfig to disable authentication for appimage
if necessary. This is just a WORKAROUND for the actual bug which is
blocked on fix from SYSFW. This is strictly temporary and should be
reverted when the fix becomes available from SYSFW.

Fixes: MCUSDK-8945
Signed-off-by: Anand Mahadevan SS <a.mahadevanss@ti.com>