am243x: xip: fix xip benchmark example in LP
The XIP benchmark example fails to boot up in AM243x-LP. This seems to
be because of cache and mpu config placed in the rodata region which goes
into the flash since text and rodata is placed in flash in XIP examples.
To correct this create a subsection in rodata and place these config
structures in that region.

Fixes: MCUSDK-9790
Signed-off-by: Anand Mahadevan SS <a.mahadevanss@ti.com>