allsoc: flash: fix the chip erase
Chip erase in the flash driver is handled by the blk/sector
erase function itself. This happens when the block number
passed is -1. But the erase timeout used is the standard
block erase. This is incorrect. In high density flashes
the timeout can be upto several minutes. And this can vary
from flash to flash. So keep the timeout as the max u32
value so that we wait as much as required for chip erase
to complete.

Additionally, the waitReady done at the start of the erase
functions seems to put the flash in a bad state in some flashes.
This is anyway a moot function call, so remove this to
avoid complexity.

Fixes: MCUSDK-10627
Signed-off-by: Anand Mahadevan SS <a.mahadevanss@ti.com>