am64x/am243x/am263x: dpl: add wfi demo example
  - add wfi demo example to demonstrate the working of wfi (standby
mode).
  - create a multicore system project in which the main core will
measure the current values with the help of INA228/IN226 on board
current monitor sensor and put the other core in wfi mode by passing a
ipc notify message.
  - check the current and power difference between the steady state and
wfi state and match with the expected power drop. If the actual values
meets the expectations, mark the test as passed.

Fixes: MCUSDK-10581
Signed-off-by: Abishek SS <abishekss@ti.com>