allsoc: sbl: move the selfcore image load
Curerntly the selfcore image load is done before other cores
start running, way before SBL ends execution. The IVT at 0x00 address
would now be of the application on self core, not of SBL. This means
that if an interrupt occurs before SBL reaches self core reset, it won't
be able to service it. To reduce this change, move the self core image
load further down.

Fixes: MCUSDK-10841
Signed-off-by: Anand Mahadevan SS <a.mahadevanss@ti.com>