allsoc: docs: update hello_world_cpp example doc
C++ linkage of libraries is not completely tested in the SDK.
Due to name mangling, sometimes it is possible that CPP built
libs might show some linker errors while building the example.
To avoid this one needs to build the libs without the CPP option
to force C linkage until the C++ support is available across the
SDK. Add this information to the documentation of the hello_world_cpp
example.

Fixes: MCUSDK-11152
Signed-off-by: Anand Mahadevan SS <a.mahadevanss@ti.com>