am243x/am64x: ospi: add option to skip flash init
In many cases, typically in an XIP application, the previous SW entity,
like the bootloader would have configured the flash in a particular
fashion. Since the application has an XIP component, doing a
reconfiguration of the flash might crash the application. To prevent
this, allow the user to specify if HW init of the flash is needed or not
in sysconfig.

Fixes: MCUSDK-10781
Signed-off-by: Anand Mahadevan SS <a.mahadevanss@ti.com>