am243x/am64x: soc: add new clock set fxn
There are certain modules for which the same parent clock can have
multiple child clocks which could be for different instances of the
module. In case of multi-core applications where there is no SW
coherence between which instances of the module are enabled, it is
possible that setting the frequency one might affect the other. To
counter this, provide a new API to reconfigure the clock with a given
parent rather than the clock set API automatically picking the first
available parent.

Fixes: MCUSDK-12117, MCUSDK-12242
Signed-off-by: Anand Mahadevan SS <a.mahadevanss@ti.com>