allsoc: ospi_flash_io: remove sector erase
Remove the sector erase from the ospi_flash_io example and replace
it with the block erase. This is because multiple erase support is
subject to the sector configuration of the flash used and we can't
take it for granted. Also mention the flash UG about the APIs and
their usage.

Fixes: MCUSDK-11667
Signed-off-by: Anand Mahadevan SS <a.mahadevanss@ti.com>