am263x/am263px: sdl: Fix DCC APIs for Intr Status and Clearing
SDL_DCC_getStatus was incorrectly checking the whether the Done
and Error Interrupts were enabled in order return the interrupt
status instead of the interrupt status register.

This patch updates SDL_DCC_getStatus to use the status register
to determine whether or not there is a Done or Error interrupt
pending.

Also, SDL_DCC_clearIntr was potentially clearing an undesiered interrupt.
This is because the register is write one to clear, but the API was first
reading the contents, updating the desired bit to 1, then writing back
to the register. If an interrupt other than the one requested to be cleared
was pending, it would also be cleared by the call.

This patch updates the SDL_DCC_clearIntr API to only clear the
requested interrupt.

Fixes: PROC_SDL-7047, PROC_SDL-7048

Signed-off-by: Angela Stegmaier <angelabaker@ti.com>