am64x/am243x: pcie: Improve EP compatibility
Initial testing with a AM64x Linux RC showed problems because Linux
wouldn't assign resources to the device if the class code was set to
0x00 (no class code / before class codes were assigned).
On x86 resources are already asigned by BIOS, so this wasn't an issue
before.
This can be fixed by using the class code 0xff (and sub class code 0xff,
although that isn't strictly necessary), which indicates a device that
doesn't fit into any of the defined classes.

Additionally the timing of the PCIe hot reset that the RC performs
before and after accessing the device via VFIO appears to be different,
with the AM64x Linux RC restoring EP configuration registers before
the EP re-initialized these registers after detecting the link.
Since this initialization step is only required once after power-up
we can avoid this race condition by performing it only once after
initial link detection.

*** Note: even with these changes, a AM64x Linux RC still wont be able
to work with the pcie_enumerate_ep example using the ti_sample_vfio test
program. Doing so requires additional changes to the ti_sample_vfio test
program that are not yet implemented. It would also be possible to use a
kernel based driver (not yet implemented). ***

Fixes: MCUSDK-12851

Signed-off-by: Dominic Rath <d-rath@ti.com>