am263px/am263x: sdfm: add dmaread example
Add DMA example to showcase the FIFO read for the DMA example. The FIFO
in the SDFM is a single register with multiple levels in HW. for
example, for each read of the FIFO DATA from a given filter, new DATA
will be popped at the same register, until the total number of the FIFO
is reached.

Fixes: MCUSDK-25,MCUSDK-4578

Signed-off-by: Mihira Madhava Bollapragada <madhava@ti.com>