am64x/am243x: sdl: fix ecc testing failures
The following changes are made to address failures in testing
for VTM aggregator, MSRAM aggregators, and PSRAM aggregator
when tested from the R5F core:

- Make MSRAM regions as non-cached so that a read of the memory
after the error injection will result in the access which is
needed to trigger the error event
- Add a RAT entry for PSRAM region at address 0x0 so that it
can be read by the R5F core in order to trigger the error event
after the error injection. Also make the memory at the translated
address as non-cached.
- Enable only one set of VTM ESM events. The VTM ECC events are
routed to both the MCU and the Main ESMs. In order to monitor all
ESM events from a single core, all the MCU ESM events are routed
to the Main ESM via the 3 MCU ESM events which are input to the
Main ESM. If the VTM ECC events are enabled for both the MCU and
the Main ESMs, then we wind up in a looping scenario. Only the
Main ESM VTM ECC events are needed to be enabled.

Fixes: PROC_SDL-6445

Signed-off-by: Angela Stegmaier <angelabaker@ti.com>