am263px: resolver : update getGain API
the getGainEstimated values is supposed to return the gain estimation
values of sin and cosine. Rather, from the MMR, this returns a value in
the range of 2^29, that could have been the max squared value of the
full range input signal.

the correct usage of the MMR is to read, then perform sqrt((2^29)/value)

The API now performs the (2^29)/value part and returns the squared gain
value. the user has to perform a sqrt operation in which ever way they
see fit, like math lib, tmu, etc.

Fixes: MCUSDK-13826

Signed-off-by: Mihira Madhava Bollapragada <madhava@ti.com>