am261x: clocktree: add variant of ZFG package
All the am261x-lp projects operate on the 500 MHz
as the ZFG package is tied to the clocktree metadata
corresponding to 500 MHz only. The clock module which
was earlier configurable from gui, now picks the
value from clocktree as read-only value. If the user
wants to change the frequency, there's no way to do so.
So we have to create another variant for am261x_zfg
to support the 400MHz also. Add the changes to
support this.

Fixes: MCUSDK-14795

Signed-off-by: Sudeshna Chaudhuri <s-chaudhuri@ti.com>