am64x/am243x: ospi: Replace sector erase with block erase
When there is no phy pattern flashed in the flash, in the flash sequence,
the phy pattern is written to the flash. Flash erase sector api is
called, since the flash is set in uniform mode, block erase api should
be called. Replace the sector erase api, with erase block api.

Fixes: SITSW-6765

Signed-off-by: Piyush Sharma <p-sharma3@ti.com>